requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
playwright>=1.40.0
pandas==2.1.4
bs4==0.0.1
aiohttp==3.9.1
python-dotenv==1.0.0